+++ packages/snap-preact-components/src/components/Molecules/Select/Select.tsx
@@ -107,7 +107,7 @@ export const Select = observer((properties: SelectProps): JSX.Element => {
 				disabled,
 			}),
 			// component theme overrides
-			...props.theme?.components?.dropdown,
+			theme: props.theme,
 		},
 		button: {
 			// default props
@@ -123,7 +123,7 @@ export const Select = observer((properties: SelectProps): JSX.Element => {
 				borderColor,
 			}),
 			// component theme overrides
-			...props.theme?.components?.button,
+			theme: props.theme,
 		},
 		icon: {
 			// default props
@@ -137,7 +137,7 @@ export const Select = observer((properties: SelectProps): JSX.Element => {
 				size: '14px',
 			}),
 			// component theme overrides
-			...props.theme?.components?.icon,
+			theme: props.theme,
 		},
 	};