+++ packages/snap-store-mobx/src/Search/SearchStore.ts
@@ -48,7 +48,7 @@ export class SearchStore extends AbstractStore {
 		this.search = new QueryStore(this.services, data.search);
 		this.facets = new FacetStore(this.config, this.services, this.storage, data.facets, data.pagination, this.meta);
 		this.filters = new FilterStore(this.services, data.filters, this.meta);
-		this.results = new ResultStore(this.services, data.results, data.pagination, data.merchandising);
+		this.results = new ResultStore(this.config, this.services, data.results, data.pagination, data.merchandising);
 		this.pagination = new PaginationStore(this.config, this.services, data.pagination);
 		this.sorting = new SortingStore(this.services, data.sorting, data.search, this.meta);
 	}